pyproject.toml
[project]
name = "Smart-Home-Assistant"
version = "None"
description = "Default template for PDM package"
authors = [
    {name = "Orb Studio", email = "cps3362521@gmail.com"},
]
dependencies = ["black>=25.1.0", "pylint>=3.3.7", "ptpython>=3.0.30", "streamlit>=1.45.1", "paho-mqtt>=2.1.0", "requests>=2.32.3", "fastapi>=0.115.12", "flask>=3.1.1", "rich>=14.0.0", "urllib3>=2.4.0", "pipreqs>=0.5.0", "passlib>=1.7.4", "sqlmodel>=0.0.24", "uvicorn>=0.34.2", "pydantic-settings>=2.9.1", "python-jose[cryptography]>=3.4.0", "python-multipart>=0.0.20", "cryptography>=44.0.3"]
requires-python = "==3.11.*"
readme = "README.md"
license = {text = "Mozilla Public License Version 2.0"}


[tool.pdm]
distribution = false
